requirements.txt
requests>=2.31
python-dotenv>=1.0
openai-whisper>=20231117
